Make a base LaTex template for creating cheat sheets
Provide a LaTex class for starting a new cheat sheet for an
application's keyboard shortcuts. Icons are provided using
FontAwesome. As an example, it creates a keyboard shortcuts for
Android studio. It pulls the shortcuts from Android's documentation
website.

There's a python script that pulls the shortcuts from the webiste.